perf: use bf16 for LoRA computation instead of fp32
Replace .float() with .to(bfloat16) in LoRA matmul across _forward_xpu,
_fused_gelu_mlp_xpu, and _compute_lora_residual_xpu. Eliminates
bf16→fp32→bf16 dtype conversion overhead.

Z-Image 512x512 9-step warm: 9.01s → 8.01s (12% faster, 1.12 it/s)

+++ src/nunchaku_torch/models/linear.py
@@ -180,8 +180,8 @@ class SVDQW4A4Linear(nn.Module):
             M, N = x_gemm.shape[0], self.out_features
             dst = torch.zeros(M, N, dtype=torch.bfloat16, device=x_gemm.device)
             if has_lora:
-                lora_out = (x.float() @ self.proj_down.float()) @ self.proj_up.float().T
-                dst.add_(lora_out.to(torch.bfloat16))
+                lora_out = (x.to(torch.bfloat16) @ self.proj_down.to(torch.bfloat16)) @ self.proj_up.to(torch.bfloat16).T
+                dst.add_(lora_out)
                 del lora_out
             if self.bias is not None:
                 dst.add_(self.bias.to(torch.bfloat16))
